requirements.txt
# Data engineering books project: load PDFs to Snowflake + Mistral/Snowflake agent.
# Install: python -m venv .venv && .venv/bin/pip install -r requirements.txt
#
# Used by:
#   scripts/load_books_to_snowflake.py  -> unstructured[pdf], snowflake
#   scripts/mistral_snowflake_agent.py   -> langchain, langchain-community, langchain-experimental, pandas
#   scripts/snowflake_helper.py          -> snowflake
#
# unstructured[pdf] pulls in unstructured-inference (and torch) for PDF layout;
# needed for partition_pdf() even with strategy="fast" due to package imports.
# For agent-only (no load_books): use requirements-agent-only.txt for a small venv.

snowflake-connector-python>=3.18
langchain>=0.3
langchain-community>=0.3
langchain-experimental>=0.3
pandas>=2.0
python-dotenv>=1.2
pypdf>=4.0
unstructured[pdf]>=0.20,<1.0
